Create connect winc1500 to wifi.ino
code for connecting the WINC1500 Xplained Pro board to a Wi-Fi network using the ASF Wi-Fi library:
This example code initializes the WINC1500 module and connects to a Wi-Fi network using the specified SSID and password. If the connection is successful, it prints a message to the console. You can add additional Wi-Fi tasks to the code as needed.
Had trouble geting it to work, will publish final version but this should guide you